Validate only relevant parts of lvm+resize_root
Create scheduler with only relevant modules to validate the scenario
and make sure that the /root is resized.
The partitioning_resize_root.pm has been modified to use the Object_Oriented
framework. Minor changes added. A function to open the current suggestion is added
and the existing _run_expert_partitioner_ funtion is adjast to make available the choice
between the latest and the option of _the existing_partition_.
I renamed also the accept_changes and created a new one with the same name to make
clear the separation of their functionality.

Condition of _existing-partition_ and corresponding constant in the _select_item_in_system_view_table_
removed as it is not used and it does not provide any usability.