Update dependencies
While looking into dependencies of several repos, I noticed that this is
missing some, for example Data::Dump, which is currently in the os-autoinst
package, but will be removed because it's not used there.

So I automatically generated the dependencies with
https://metacpan.org/pod/Perl::PrereqScanner::NotQuiteLite

    scan-perl-prereqs-nqlite --cpanfile >cpanfile.new

and copied most of them to the existing cpanfile.
We can't use the autogenerated cpanfile directly though, because it
contains also os-autoinst modules (testapi, ...).

scan-perl-prereqs-nqlite also provides a JSON output, so maybe this can
be used in the future to have more automation.