Create autoyast profiles and schedulers for HPC
ay profiles are generated with the clone.pm module and were adjusted in order
to remove unnecessary elements and modified to run in each sle version. The
main difference between the two xml are the packages, the modules and the
addons in the suse_register tag. Other than that the product had to be adjust
as well from SLES to SLE_HPC to make HPC module available.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>